ai_service/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.0
langchain==0.2.0
langgraph==0.1.19
groq==0.9.0
paddlepaddle==2.6.1
paddleocr==2.7.3
pdf2image==1.17.0
insightface==0.7.3
onnxruntime==1.18.0
opencv-python-headless==4.10.0.84
sentence-transformers==3.0.0
qdrant-client==1.9.0
minio==7.2.7
aiokafka==0.10.0
httpx==0.27.0
redis==5.0.4
pydantic==2.7.0
pydantic-settings==2.3.0
python-dotenv==1.0.1
numpy==1.26.4
Pillow==10.3.0
scikit-learn==1.5.0
xgboost==2.0.3
transformers==4.41.0
torch==2.3.0
pytest==8.2.0
pytest-asyncio==0.23.7
aiofiles==23.2.1
python-multipart==0.0.9
